src/backend/requirements.txt
fastapi
uvicorn
google-genai
requests
beautifulsoup4
python-docx
google-generativeai
pandas